# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kombucha Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  clerkUserId: string;
  hasCompletedOnboarding: boolean;
  hasMadeBefore: boolean;
  hasScoby: boolean;
  currentStage?: string | null;
  experienceLevel?: string | null;
  flavorPreference?: string | null;
  onboardingAdvice?: string | null;
  ttsEnabled: boolean;
  createdAt: string;
}

export interface UpdateProfileBody {
  hasCompletedOnboarding?: boolean;
  hasMadeBefore?: boolean;
  hasScoby?: boolean;
  currentStage?: string;
  experienceLevel?: string;
  flavorPreference?: string;
  onboardingAdvice?: string;
  ttsEnabled?: boolean;
}

export type BatchStatus = (typeof BatchStatus)[keyof typeof BatchStatus];

export const BatchStatus = {
  active: "active",
  f1_complete: "f1_complete",
  f2_complete: "f2_complete",
  archived: "archived",
} as const;

export interface Batch {
  id: number;
  userId: string;
  name: string;
  startedAt: string;
  status: BatchStatus;
  teaType?: string | null;
  notes?: string | null;
  logCount: number;
  photoCount: number;
  daysSinceStart: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBatchBody {
  name: string;
  startedAt?: string;
  teaType?: string;
  notes?: string;
}

export type UpdateBatchBodyStatus =
  (typeof UpdateBatchBodyStatus)[keyof typeof UpdateBatchBodyStatus];

export const UpdateBatchBodyStatus = {
  active: "active",
  f1_complete: "f1_complete",
  f2_complete: "f2_complete",
  archived: "archived",
} as const;

export interface UpdateBatchBody {
  name?: string;
  status?: UpdateBatchBodyStatus;
  teaType?: string;
  notes?: string;
}

export interface Log {
  id: number;
  batchId: number;
  dayNumber: number;
  temperature?: number | null;
  scobylook?: string | null;
  smell?: string | null;
  color?: string | null;
  notes?: string | null;
  aiTip?: string | null;
  loggedAt: string;
  createdAt: string;
}

export type CreateLogBodySmell =
  (typeof CreateLogBodySmell)[keyof typeof CreateLogBodySmell];

export const CreateLogBodySmell = {
  good: "good",
  sour: "sour",
  vinegary: "vinegary",
  strange: "strange",
} as const;

export interface CreateLogBody {
  dayNumber: number;
  temperature?: number;
  scobylook?: string;
  smell?: CreateLogBodySmell;
  color?: string;
  notes?: string;
  loggedAt?: string;
}

export interface Photo {
  id: number;
  batchId: number;
  objectPath: string;
  caption?: string | null;
  dayNumber?: number | null;
  aiAnalysis?: string | null;
  takenAt: string;
  createdAt: string;
}

export interface CreatePhotoBody {
  objectPath: string;
  caption?: string;
  dayNumber?: number;
  takenAt?: string;
  analyzeWithAi?: boolean;
}

export interface AnalyzePhotoBody {
  objectPath: string;
}

export interface AnalyzePhotoResponse {
  analysis: string;
}

export interface OnboardingAdviceBody {
  hasMadeBefore: boolean;
  hasScoby: boolean;
  currentStage?: string;
  experienceLevel?: string;
}

export interface OnboardingAdviceResponse {
  advice: string;
}

export interface ChatMessageBody {
  message: string;
}

export interface ChatMessageResponse {
  reply: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface TtsBody {
  text: string;
}

export interface TtsResponse {
  /** Base64 encoded audio (mp3) */
  audio: string;
  format: string;
}

export type FlavoringGuideResponseSuggestionsItem = {
  name: string;
  ingredients: string;
  tip: string;
};

export interface FlavoringGuideResponse {
  suggestions: FlavoringGuideResponseSuggestionsItem[];
  generalTips: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface DashboardSummary {
  activeBatchCount: number;
  totalBatchCount: number;
  totalPhotoCount: number;
  recentLogs: Log[];
  activeBatches: Batch[];
}

export type GetFlavoringGuideParams = {
  preference?: GetFlavoringGuidePreference;
};

export type GetFlavoringGuidePreference =
  (typeof GetFlavoringGuidePreference)[keyof typeof GetFlavoringGuidePreference];

export const GetFlavoringGuidePreference = {
  sweet: "sweet",
  sour: "sour",
  fruity: "fruity",
} as const;