# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kombucha Tracker API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      summary: List all fermentation batches for the current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
        "401":
          description: Unauthorized
    post:
      operationId: createBatch
      tags: [batches]
      summary: Create a new fermentation batch
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchBody"
      responses:
        "201":
          description: Created batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "401":
          description: Unauthorized
  /batches/{batchId}:
    get:
      operationId: getBatch
      tags: [batches]
      summary: Get a single batch by ID
      security:
        - clerkAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batch details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    put:
      operationId: updateBatch
      tags: [batches]
      summary: Update a batch (e.g., mark F1 complete)
      security:
        - clerkAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBatchBody"
      responses:
        "200":
          description: Updated batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteBatch
      tags: [batches]
      summary: Delete a batch
      security:
        - clerkAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
...
  /ai/analyze-photo:
    post:
      operationId: analyzePhoto
      tags: [ai]
      summary: Analyze a SCOBY/batch photo with AI
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzePhotoBody"
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzePhotoResponse"
        "401":
          description: Unauthorized
  /ai/onboarding-advice:
    post:
      operationId: getOnboardingAdvice
      tags: [ai]
      summary: Get personalized AI advice after onboarding quiz
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingAdviceBody"
      responses:
        "200":
          description: AI onboarding advice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingAdviceResponse"
        "401":
          description: Unauthorized
  /ai/chat:
    post:
      operationId: chatWithAdvisor
      tags: [ai]
      summary: Chat with the kombucha AI advisor
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageBody"
      responses:
        "200":
          description: AI chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageResponse"
        "401":
          description: Unauthorized
  /ai/chat/history:
    get:
      operationId: getChatHistory
      tags: [ai]
      summary: Get chat conversation history for the current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "401":
          description: Unauthorized
  /ai/tts:
    post:
      operationId: textToSpeech
      tags: [ai]
      summary: Convert text to speech audio (base64)
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TtsBody"
      responses:
        "200":
          description: Audio as base64
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TtsResponse"
        "401":
          description: Unauthorized
  /ai/flavoring-guide:
    get:
      operationId: getFlavoringGuide
      tags: [ai]
      summary: Get flavoring suggestions when F1 is complete
      security:
        - clerkAuth: []
      parameters:
        - name: preference
          in: query
          required: false
          schema:
            type: string
            enum: [sweet, sour, fruity]
      responses:
        "200":
          description: Flavoring suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlavoringGuideResponse"
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Batch:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        startedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, f1_complete, f2_complete, archived]
        teaType:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        logCount:
          type: integer
        photoCount:
          type: integer
        daysSinceStart:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - startedAt
        - status
        - logCount
        - photoCount
        - daysSinceStart
        - createdAt
        - updatedAt
    CreateBatchBody:
      type: object
      properties:
        name:
          type: string
        startedAt:
          type: string
          format: date-time
        teaType:
          type: string
        notes:
          type: string
      required:
        - name
    UpdateBatchBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, f1_complete, f2_complete, archived]
        teaType:
          type: string
        notes:
          type: string
...
    AnalyzePhotoBody:
      type: object
      properties:
        objectPath:
          type: string
      required:
        - objectPath
    AnalyzePhotoResponse:
      type: object
      properties:
        analysis:
          type: string
      required:
        - analysis
    OnboardingAdviceBody:
      type: object
      properties:
        hasMadeBefore:
          type: boolean
        hasScoby:
          type: boolean
        currentStage:
          type: string
        experienceLevel:
          type: string
      required:
        - hasMadeBefore
        - hasScoby
    OnboardingAdviceResponse:
      type: object
      properties:
        advice:
          type: string
      required:
        - advice
    ChatMessageBody:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ChatMessageResponse:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - content
        - createdAt
    TtsBody:
      type: object
      properties:
        text:
          type: string
      required:
        - text
    TtsResponse:
      type: object
      properties:
        audio:
          type: string
          description: Base64 encoded audio (mp3)
        format:
          type: string
      required:
        - audio
        - format
    FlavoringGuideResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              ingredients:
                type: string
              tip:
                type: string
            required:
              - name
              - ingredients
              - tip
        generalTips:
          type: string
      required:
        - suggestions
        - generalTips
...