# lib/api-zod/src/generated/types/analyzePhotoResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kombucha Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzePhotoResponse {
  analysis: string;
}