# lib/api-zod/src/generated/types/updateBatchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kombucha Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBatchBodyStatus } from "./updateBatchBodyStatus";

export interface UpdateBatchBody {
  name?: string;
  status?: UpdateBatchBodyStatus;
  teaType?: string;
  notes?: string;
}